reorganized resolvers, fixed session bug
